fix syncing to qa02 evm

+++ tools/evmtest/node_init_qa02.sh
@@ -5,12 +5,14 @@ NAMESPACE=qa02
 SERV_URL=https://${ENV}-${NAMESPACE}.${ENV}.findora.org
 FINDORAD_IMG=public.ecr.aws/k6m5b6e2/release/findorad:testnet-evm-v0.2.7a
 
-ROOT_DIR=$1
+ROOT_DIR=$2
 [ -n "$ROOT_DIR" ] || ROOT_DIR=/data/findora/$NAMESPACE
 
-FRESH_START=1
-if [ "$1" = "snapshot" ];then
-    FRESH_START=0
+START_MODE=1
+if [ "$1" = "snapshot" ]; then
+    START_MODE=0
+elif [ "$1" = "restart" ]; then
+  START_MODE=2
 else
     sudo rm -rf \
     "$ROOT_DIR"/bin \
@@ -21,32 +23,34 @@ fi
 
 
 check_env() {
-    for i in wget curl; do
-        which $i >/dev/null 2>&1
-        if [[ 0 -ne $? ]]; then
-            echo -e "\n\033[31;01m${i}\033[00m has not been installed properly!\n"
-            exit 1
-        fi
-    done
-
-    if ! [ -f "$keypath" ]; then
-        echo "The key file doesnot exist: $keypath"
-        exit 1
+  ((START_MODE != 2)) || return 1
+  for i in wget curl; do
+    which $i >/dev/null 2>&1
+    if [[ 0 -ne $? ]]; then
+      echo -e "\n\033[31;01m${i}\033[00m has not been installed properly!\n"
+      exit 1
     fi
+  done
+
+  if ! [ -f "$keypath" ]; then
+    echo "The key file doesnot exist: $keypath"
+    exit 1
+  fi
 }
 
 set_binaries() {
-    OS=$1
+  ((START_MODE != 2)) || return 1
+  OS=$1
 
-    docker pull ${FINDORAD_IMG} || exit 1
-    wget -T 10 https://wiki.findora.org/bin/${OS}/fn || exit 1
+  docker pull ${FINDORAD_IMG} || exit 1
+  wget -T 10 https://wiki.findora.org/bin/"$OS"/fn || exit 1
 
-    new_path=${ROOT_DIR}/bin
+  new_path=${ROOT_DIR}/bin
 
-    rm -rf $new_path 2>/dev/null
-    mkdir -p $new_path || exit 1
-    mv fn $new_path || exit 1
-    chmod -R +x ${new_path} || exit 1
+  rm -rf $new_path 2>/dev/null
+  mkdir -p $new_path || exit 1
+  mv fn $new_path || exit 1
+  chmod -R +x ${new_path} || exit 1
 }
 
 keypath=${ROOT_DIR}/${NAMESPACE}_node.key
@@ -54,38 +58,37 @@ FN=${ROOT_DIR}/bin/fn
 
 check_env
 
-if [[ "Linux" == `uname -s` ]]; then
-    set_binaries linux
-# elif [[ "FreeBSD" == `uname -s` ]]; then
-    # set_binaries freebsd
-elif [[ "Darwin" == `uname -s` ]]; then
-    set_binaries macos
+if [[ "Linux" == "$(uname -s)" ]]; then
+  set_binaries linux
+elif [[ "Darwin" == "$(uname -s)" ]]; then
+  set_binaries macos
 else
-    echo "Unsupported system platform!"
-    exit 1
+  echo "Unsupported system platform!"
+  exit 1
 fi
 
 ######################
 # Config local node #
 ######################
+if ((START_MODE != 2)); then
+  node_mnemonic=$(grep 'Mnemonic' "$keypath" | sed 's/^.*Mnemonic:[^ ]* //')
+  #xfr_pubkey="$(grep 'pub_key' "$keypath" | sed 's/[",]//g' | sed 's/ *pub_key: *//')"
 
-node_mnemonic=$(grep 'Mnemonic' "$keypath" | sed 's/^.*Mnemonic:[^ ]* //')
-#xfr_pubkey="$(grep 'pub_key' "$keypath" | sed 's/[",]//g' | sed 's/ *pub_key: *//')"
+  echo "$node_mnemonic" > ${ROOT_DIR}/node.mnemonic || exit 1
 
-echo "$node_mnemonic" > ${ROOT_DIR}/node.mnemonic || exit 1
+  $FN setup -S ${SERV_URL} || exit 1
+  $FN setup -K ${ROOT_DIR}/tendermint/config/priv_validator_key.json || exit 1
+  $FN setup -O ${ROOT_DIR}/node.mnemonic || exit 1
 
-$FN setup -S ${SERV_URL} || exit 1
-$FN setup -K ${ROOT_DIR}/tendermint/config/priv_validator_key.json || exit 1
-$FN setup -O ${ROOT_DIR}/node.mnemonic || exit 1
+  # clean old data and config files
+  sudo rm -rf ${ROOT_DIR}/findorad || exit 1
 
-# clean old data and config files
-sudo rm -rf ${ROOT_DIR}/findorad || exit 1
+  docker run --rm -v ${ROOT_DIR}/tendermint:/root/.tendermint ${FINDORAD_IMG} init --${NAMESPACE} || exit 1
 
-docker run --rm -v ${ROOT_DIR}/tendermint:/root/.tendermint ${FINDORAD_IMG} init --${NAMESPACE} || exit 1
+  sudo chown -R "$(id -u)":"$(id -g)" ${ROOT_DIR}/tendermint/
+fi
 
-sudo chown -R "$(id -u)":"$(id -g)" ${ROOT_DIR}/tendermint/
-
-if ((FRESH_START == 1)); then
+if ((START_MODE == 1)); then
     mkdir -p ${ROOT_DIR}/tendermint/config
     mkdir -p ${ROOT_DIR}/tendermint/data
     
@@ -94,7 +97,7 @@ if ((FRESH_START == 1)); then
 
     rm -rf "${ROOT_DIR}/findorad"
     rm -rf "${ROOT_DIR}/tendermint/config/addrbook.json"
-else
+elif ((START_MODE == 0)); then
     ###################
     # get snapshot    #
     ###################
@@ -130,8 +133,10 @@ docker run -d \
     -p 8668:8668 \
     -p 8667:8667 \
     -p 26657:26657 \
+    -p 8545:8545 \
+    -e EVM_CHAIN_ID=516 \
     --name findorad \
-    ${FINDORAD_IMG} node\
+    ${FINDORAD_IMG} node \
     --ledger-dir /tmp/findora \
     --tendermint-host 0.0.0.0 \
     --tendermint-node-key-config-path="/root/.tendermint/config/priv_validator_key.json" \
